Migrate getNextHour to getNextHourInstant.
This migrates getNextHour to use the java.time API for underlying
time handling logic. The deprecated function is renamed to getNextHourDate
while its replacement is named as getNextHourInstant to disambiguate.

getNextHourInstant calls getNextHourInstantFromNow. The latter is
for testing purposes since it exploits a quirk in how java.time.Instant
can only be rounded down using truncatedTo.