Migrate StringHelper.toUtcFormat to TimeHelper.
This moves StringHelper.toUtcFormat to
TimeHelper.formatTimeZoneToUtcOffset and makes the underlying logic
clearer using java.time.ZoneOffset. Since this method expects a
timezone offset as a double, it is marked deprecated and slated for
removal as we migrate to java.time completely.

Added tests to cover TimeHelper.formatTimeZoneToUtcOffset so that
its behavior will be clear when completing migration.